Fixed typo and moved type declaration
