+++ src/app/admin/experiences/create/page.tsx
@@ -1,4 +1,4 @@
-import ExperiencesForm from "@/components/ExperiencesForm/EsperiencesForm";
+import ExperiencesForm from "@/components/ExperiencesForm/ExperiencesForm";
 
 const page = () => {
   return (
