+++ src/components/ExperiencesForm/ExperiencesForm.types.ts
@@ -1,4 +1,7 @@
-import { ExperiencesFormData } from "./ExperiencesForm.schema";
+import { z } from "zod";
+import { ExperiencesFormSchema } from "./ExperiencesForm.schema";
+
+export type ExperiencesFormData = z.infer<typeof ExperiencesFormSchema>;
 
 export type ExperiencesFormParams = {
   initialValues?: ExperiencesFormData;
